Adding 2nd recipes in my odin-recipes website.
Gajar ka Halwa is the second recipes in my website. because Gajar Halwa holds a special place in Indian cuisine and is often prepared during festivals like Diwali and Holi, as well as during winter months when carrots are in season. Its rich, creamy texture and aromatic flavor make it a beloved dessert that is enjoyed by people of all ages