Adding 3rd recipes in my odin-recipes website.
MalaiKulfi is my Third recipes in my website. beacause "Malai Kulfi" is a type of frozen dessert from India. It's like ice cream but creamier and richer. It's made by boiling milk with sugar and flavorings like cardamom and nuts, then freezing it in molds until it's solid. It's a delicious treat enjoyed especially in hot weather.